remove circular dependencies + renaming of components + layout fix
